Laser now starts at camera's location
I don’t think that the space ship location is ever updating. For now,
the laser begins at camera location. I know that this isn’t ideal but
it works for now. If you want, you can start adding the physics to the
lasers. They are spheres. I plan on making them smaller later on.